Move express server script to frontend root
